refactor(imgproc): hoist residency dispatch out of color into cuda::dispatch
Phase 1 of the CUDA-integration plan — no behavior change. Residency,
DeviceExec (with its cross-stream event-fence protocol), is_device,
pair_residency, device_exec_for, untyped_device_err, and the
device_slices! macro move from color/cuda_dispatch.rs to a new
cuda/dispatch.rs, so the upcoming geometry device paths (resize, warps)
share one copy of the machinery instead of reimplementing it.
color/cuda_dispatch.rs keeps only what is elementwise-specific: the size
check and the adapter macros over the color launchers. The existing color
CUDA tests are the regression guard — all green on Orin.
